+++ src/utils/extensions.ts
@@ -1,8 +1,10 @@
 import * as vscode from "vscode";
 import * as https from "https";
 import { IncomingMessage } from "http";
-import { checkQuartoVersion, installQuartoExtension } from "./quarto";
 import * as crypto from "crypto";
+import * as fs from "fs";
+import * as path from "path";
+import { installQuartoExtension } from "./quarto";
 
 interface ExtensionQuickPickItem extends vscode.QuickPickItem {
 	url?: string;
@@ -75,35 +77,53 @@ export async function installQuartoExtensions(
 	selectedExtensions: readonly ExtensionQuickPickItem[],
 	log: vscode.OutputChannel
 ) {
+	const workspaceFolder = vscode.workspace.workspaceFolders?.[0].uri.fsPath || "";
 	const mutableSelectedExtensions: ExtensionQuickPickItem[] = [...selectedExtensions];
 
-	const isQuartoAvailable = await checkQuartoVersion();
-	if (!isQuartoAvailable) {
-		const message =
-			"Quarto is not installed or not available in PATH. Please install Quarto and make sure it is available in PATH.";
-		log.appendLine(message);
-		vscode.window.showErrorMessage(message);
-		return;
+	const config = vscode.workspace.getConfiguration("quartoWizard.ask");
+	let configTrustAuthors = config.get<string>("trustAuthors");
+	let configConfirmInstall = config.get<string>("confirmInstall");
+
+	if (configTrustAuthors === "always") {
+		const trustAuthors = await vscode.window.showQuickPick(
+			[
+				{ label: "Yes", description: "Trust authors." },
+				{ label: "No", description: "Do not trust authors." },
+				{ label: "Never ask again", description: "Change setting to never ask again." },
+			],
+			{
+				placeHolder: "Do you trust the authors of the selected extension(s)?",
+			}
+		);
+		if (trustAuthors?.label === "Never ask again") {
+			await config.update("trustAuthors", "never", vscode.ConfigurationTarget.Global);
+		} else if (trustAuthors?.label !== "Yes") {
+			const message = "Operation cancelled because the authors are not trusted.";
+			log.appendLine(message);
+			vscode.window.showInformationMessage(message);
+			return;
+		}
 	}
 
-	const trustAuthors = await vscode.window.showQuickPick(["Yes", "No"], {
-		placeHolder: "Do you trust the authors of the selected extension(s)?",
-	});
-	if (trustAuthors !== "Yes") {
-		const message = "Operation cancelled because the authors are not trusted.";
-		log.appendLine(message);
-		vscode.window.showInformationMessage(message);
-		return;
-	}
-
-	const installWorkspace = await vscode.window.showQuickPick(["Yes", "No"], {
-		placeHolder: "Do you want to install the selected extension(s)?",
-	});
-	if (installWorkspace !== "Yes") {
-		const message = "Operation cancelled by the user.";
-		log.appendLine(message);
-		vscode.window.showInformationMessage(message);
-		return;
+	if (configConfirmInstall === "always") {
+		const installWorkspace = await vscode.window.showQuickPick(
+			[
+				{ label: "Yes", description: "Install extensions." },
+				{ label: "No", description: "Do not install extensions." },
+				{ label: "Never ask again", description: "Change setting to never ask again." },
+			],
+			{
+				placeHolder: "Do you want to install the selected extension(s)?",
+			}
+		);
+		if (installWorkspace?.label === "Never ask again") {
+			await config.update("confirmInstall", "never", vscode.ConfigurationTarget.Global);
+		} else if (installWorkspace?.label !== "Yes") {
+			const message = "Operation cancelled by the user.";
+			log.appendLine(message);
+			vscode.window.showInformationMessage(message);
+			return;
+		}
 	}
 
 	vscode.window.withProgress(
@@ -130,8 +150,12 @@ export async function installQuartoExtensions(
 				}
 				progress.report({
 					message: `(${installedCount} / ${totalExtensions}) ${selectedExtension.label} ...`,
-					increment: (1 / totalExtensions) * 100,
+					increment: (1 / (totalExtensions + 1)) * 100,
 				});
+
+				const extensionsDirectory = path.join(workspaceFolder, "_extensions");
+				const existingExtensions = getMtimeExtensions(extensionsDirectory);
+
 				const success = await installQuartoExtension(selectedExtension.description, log);
 				if (success) {
 					installedExtensions.push(selectedExtension.description);
@@ -139,8 +163,29 @@ export async function installQuartoExtensions(
 					failedExtensions.push(selectedExtension.description);
 				}
 
+				// Update _extension.yml file with source, i.e., GitHub username/repo
+				// This is needed for the extension to be updated in the future
+				// To be removed when Quarto supports source records in the _extension.yml file or elsewhere
+				// See https://github.com/quarto-dev/quarto-cli/issues/11468
+				const newExtension = findModifiedExtensions(existingExtensions, extensionsDirectory);
+				const fileNames = ["_extension.yml", "_extension.yaml"];
+				const filePath = fileNames
+					.map((name) => path.join(extensionsDirectory, ...newExtension, name))
+					.find((fullPath) => fs.existsSync(fullPath));
+				if (filePath) {
+					const fileContent = fs.readFileSync(filePath, "utf-8");
+					const updatedContent = fileContent.includes("source: ")
+						? fileContent.replace(/source: .*/, `source: ${selectedExtension.description}`)
+						: `${fileContent.trim()}\nsource: ${selectedExtension.description}`;
+					fs.writeFileSync(filePath, updatedContent);
+				}
+
 				installedCount++;
 			}
+			progress.report({
+				message: `(${totalExtensions} / ${totalExtensions}) extensions processed.`,
+				increment: (1 / (totalExtensions + 1)) * 100,
+			});
 
 			if (installedExtensions.length > 0) {
 				log.appendLine(`\n\nSuccessfully installed extension${installedExtensions.length > 1 ? "s" : ""}:`);
@@ -173,4 +218,51 @@ export async function installQuartoExtensions(
 			}
 		}
 	);
+}
+
+function findQuartoExtensionsRecurse(dir: string): string[] {
+	let results: string[] = [];
+	const list = fs.readdirSync(dir);
+	list.forEach((file) => {
+		const filePath = path.join(dir, file);
+		const stat = fs.statSync(filePath);
+		if (stat && stat.isDirectory() && path.basename(filePath) !== "_extensions") {
+			results = results.concat(findQuartoExtensionsRecurse(filePath));
+		} else if (file.endsWith("_extension.yml") || file.endsWith("_extension.yaml")) {
+			results.push(filePath);
+		}
+	});
+	return results;
+}
+
+export function findQuartoExtensions(dir: string): string[] {
+	return findQuartoExtensionsRecurse(dir).map((filePath) => path.relative(dir, path.dirname(filePath)));
+}
+
+function getMtimeExtensions(dir: string): { [key: string]: Date } {
+	if (!fs.existsSync(dir)) {
+		return {};
+	}
+	const extensions = findQuartoExtensions(dir);
+	const extensionsMtimeDict: { [key: string]: Date } = {};
+	extensions.forEach((extension) => {
+		extensionsMtimeDict[extension] = fs.statSync(path.join(dir, extension)).mtime;
+	});
+	return extensionsMtimeDict;
+}
+
+function findModifiedExtensions(extensions: { [key: string]: Date }, dir: string): string[] {
+	if (!fs.existsSync(dir)) {
+		return [];
+	}
+	const modifiedExtensions: string[] = [];
+	const currentExtensions = findQuartoExtensions(dir);
+	currentExtensions.forEach((extension) => {
+		const extensionPath = path.join(dir, extension);
+		const extensionMtime = fs.statSync(extensionPath).mtime;
+		if (!extensions[extension] || extensions[extension] < extensionMtime) {
+			modifiedExtensions.push(extension);
+		}
+	});
+	return modifiedExtensions;
 }